Add working user estimate saving,  better visual structure and visual effects for button hover.

+++ app/src/main/java/studies/drawingapp/ImageComparisonResult.java
@@ -3,13 +3,13 @@ package studies.drawingapp;
 public class ImageComparisonResult {
     public final double squareError;
     public final double systemEstimate;
-    public final int usedDiffConstant;
+    public final double usedDiffConstant;
     public final double[] drawingMean;
     public final double[] modelMean;
     public final double[] drawingStdDevScale;
     public final double[] cmaesTransformations;
 
-    public ImageComparisonResult(double squareError, double systemEstimate, int usedDiffConstant,
+    public ImageComparisonResult(double squareError, double systemEstimate, double usedDiffConstant,
                                  double[] drawingMean, double[] modelMean, double[] drawingStdDevScale,
                                  double[] cmaesTransformations) {
         this.squareError = squareError;
